+++ recipes/repositories.rb
@@ -29,7 +29,7 @@ branches.each do |branch|
     if (branch == 'stable' && !node['opennms']['stable']) ||
        ((branch == 'snapshot' || branch == 'obsolete' || branch == 'oldstable') && node['opennms']['stable'])
       skip = true
-    elsif (branch == 'oldstable' && !node['opennms']['stable'])
+    elsif branch == 'oldstable' && !node['opennms']['stable']
       skip = false
     end
     # next if skip