+++ test/fixtures/cookbooks/opennms_resource_tests/recipes/import_node.rb
@@ -20,7 +20,7 @@ opennms_import_node 'nodeB' do
   categories %w(Servers Test)
   assets 'vendorPhone' => '411', 'serialNumber' => 'SN12838931'
   meta_data [{ 'context' => 'foo', 'key' => 'bar', 'value' => 'baz' }, { 'context' => 'foofoo', 'key' => 'barbar', 'value' => 'bazbaz' }]
-  sync_import false
+  sync_import true
 end
 
 node_c_foreign_id = 'nodeC'
@@ -36,11 +36,7 @@ opennms_import_node 'nodeC' do
   categories %w(Servers Test)
   assets 'vendorPhone' => '511', 'serialNumber' => 'SN12838932'
   meta_data [{ 'context' => 'foo', 'key' => 'bar', 'value' => 'baz' }, { 'context' => 'foofoo', 'key' => 'barbar', 'value' => 'bazbaz' }]
-  sync_import false
+  sync_import true
   sync_wait_periods 30
   sync_wait_secs 10
-end
-
-log 'sync dry-source after adding two nodes' do
-  notifies :sync, 'opennms_import[dry-source]', :immediately
 end